Action Key Error Fix
Some actions in odoo do not return a result key. For example the action_validate in the stock.inventory model. This causes a key error when result["result"] is returned. The line I added will assign a value of None to the result key when the result key does not exist. This fixes the key error.